Replace this.rfStorages with this.rfStorBasic and this.rfStorDense  (advanced, but string sortable/better self documenting)
This should resolve the following bug reports:
 https://github.com/McJtyMods/EFab/issues/15
 https://github.com/FTBTeam/FTB-Continuum/issues/330

As a side note, I believe the state per block multiblock design
is a mistake.  It does not scale up well and probably causes
many TE updates to occur per tick.

NOTE: Untested as the changes are minimal and setting up a build env seems daunting.

Legal:
If possible I disclaim any copyright to the code I am contributing.
I do not believe this input to be sufficient work to constitute a new copywritten work.
If the above is not possible, I assign any copyright claims for my contribution under
dual-license as both CC0 and a perpetual grant of full-non-exclusive rights to any developer of this mod.